requirements.txt
ansi2html==1.7.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
bleach==4.1.0
Brotli==1.0.9
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
dash==2.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
Flask==2.0.2
Flask-Compress==1.10.1
ghostscript==0.7
idna==3.3
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-dash==0.4.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.1
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pdfminer.six==20211012
pdftopng==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
plotly==5.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
PyPDF2==1.26.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
retrying==1.3.3
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tabulate==0.8.9
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.7.0
